refactored all instances of placeholder app name "thank u, next" to "FlipPhone". hopefully it doesn't break everything..
