+++ .idea/.name
@@ -1,1 +1,1 @@
-thank u, next
+FlipPhone
