+++ app/src/androidTest/java/com/example/flipphone/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.thankunext;
+package com.example.flipphone;
 
 import android.content.Context;
 
@@ -22,6 +22,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
 
-        assertEquals("com.example.thankunext", appContext.getPackageName());
+        assertEquals("com.example.flipphone", appContext.getPackageName());
     }
 }
