+++ app/src/main/java/com/example/flipphone/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.thankunext;
+package com.example.flipphone;
 
 import android.os.Bundle;
