+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">thank u, next</string>
+    <string name="app_name">FlipPhone</string>
     <string name="action_settings">Settings</string>
 </resources>
