+++ app/src/test/java/com/example/flipphone/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.thankunext;
+package com.example.flipphone;
 
 import org.junit.Test;
