+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name='thank u, next'
+rootProject.name='FlipPhone'
